refactor(rtc-topology): separate no-rtt tree ownership
